// apps/backend/src/appsync/API.ts
export enum VerificationStatus {
  UNCHECKED = 'UNCHECKED',
  PASS = 'PASS',
  FAIL = 'FAIL',
  PASS_MANUAL = 'PASS_MANUAL',
  FAIL_MANUAL = 'FAIL_MANUAL',
  REFER = 'REFER',
  WAIT = 'WAIT',
  ARCHIVED = 'ARCHIVED',
  INACTIVE = 'INACTIVE',
}
// ...
export enum EntityType {
  COMPANY = 'COMPANY',
  TRUST = 'TRUST',
  NOT_FOR_PROFIT = 'NOT_FOR_PROFIT',
  SELF_MANAGED_SUPER_FUND = 'SELF_MANAGED_SUPER_FUND',
  BPAY = 'BPAY',
}
// ...
export enum S3UploadLevel {
  public = 'public',
  protected = 'protected',
  private = 'private',
}

export type EntityUser = {
  __typename: 'EntityUser';
  id: string;
  entityId: string;
  userId: string;
  invitedEmail?: string | null;
  referredBy?: string | null;
  invitedEntityId?: string | null;
  firmEntityId?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  role?: EntityUserRole | null;
  paymentsEnabled?: boolean | null;
  entitySearchName?: string | null;
  entity?: Entity | null;
  searchName?: string | null;
  createdBy?: string | null;
  createdAt?: string | null;
  updatedAt?: string | null;
  status?: EntityUserStatus | null;
};

export enum EntityUserRole {
  OWNER = 'OWNER',
  ADMIN = 'ADMIN',
  ACCOUNTANT = 'ACCOUNTANT',
  CONTRIBUTOR = 'CONTRIBUTOR',
}

export type Entity = {
  __typename: 'Entity';
  id: string;
  type: EntityType;
  taxNumber?: string | null;
  companyNumber?: string | null;
  billerCode?: number | null;
  name: string;
  legalName?: string | null;
  searchName?: string | null;
  address?: Address | null;
  logo?: Image | null;
  entityBeneficialOwners?: ModelEntityBeneficialOwnerConnection | null;
  entityUsers?: ModelEntityUserConnection | null;
  gstRegistered?: boolean | null;
  providerEntityId?: string | null;
  providerCompanyId?: string | null;
  providerBillUserCompanyId?: string | null;
  providerBankAccountId?: string | null;
  providerDigitalWalletId?: string | null;
  providerBpayCrn?: string | null;
  contact?: EntityContact | null;
  paymentMethods?: ModelPaymentMethodConnection | null;
  paymentMethodId?: string | null;
  paymentUserId?: string | null;
  disbursementMethodId?: string | null;
  receivingAccounts?: ModelPaymentMethodConnection | null;
  ubosCreated?: boolean | null;
  numUbosCreated?: number | null;
  subCategory?: string | null;
  clientsStatus?: EntityClientsStatus | null;
  ocrEmail?: string | null;
  verificationStatus?: VerificationStatus | null;
  createdAt?: string | null;
  updatedAt?: string | null;
  owner: string;
  referredBy?: string | null;
};

export type Address = {
  __typename: 'Address';
  addressId?: string | null;
  placeId?: string | null;
  contactName?: string | null;
  contactNumber?: string | null;
  address1?: string | null;
  unitNumber?: string | null;
  streetNumber?: string | null;
  streetName: string;
  streetType?: string | null;
  city: string;
  country: string;
  countryCode?: string | null;
  state: string;
  stateCode?: string | null;
  postalCode: string;
};

export type Image = {
  __typename: 'Image';
  alt?: string | null;
  identityId?: string | null;
  key: string;
  level: S3UploadLevel;
  type?: string | null;
};

export type ModelEntityBeneficialOwnerConnection = {
  __typename: 'ModelEntityBeneficialOwnerConnection';
  items?: Array<EntityBeneficialOwner | null> | null;
  nextToken?: string | null;
};

export type EntityBeneficialOwner = {
  __typename: 'EntityBeneficialOwner';
  entityId: string;
  beneficialOwnerId: string;
  beneficialOwner?: BeneficialOwner | null;
  createdAt?: string | null;
  updatedAt?: string | null;
  owner?: string | null;
};

export type ModelEntityUserConnection = {
  __typename: 'ModelEntityUserConnection';
  items: Array<EntityUser | null>;
  nextToken?: string | null;
};

export type EntityContact = {
  __typename: 'EntityContact';
  firstName?: string | null;
  lastName?: string | null;
  email?: string | null;
  phone?: string | null;
  role?: string | null;
};

export type ModelPaymentMethodConnection = {
  __typename: 'ModelPaymentMethodConnection';
  items: Array<PaymentMethod | null>;
  nextToken?: string | null;
};

export type PaymentMethod = {
  __typename: 'PaymentMethod';
  id: string;
  paymentMethodType: PaymentMethodType;
  type?: string | null;
  fullName?: string | null;
  number?: string | null;
  expMonth?: string | null;
  expYear?: string | null;
  accountName?: string | null;
  bankName?: string | null;
  accountNumber?: string | null;
  routingNumber?: string | null;
  holderType?: BankHolderType | null;
  accountType?: BankAccountType | null;
  status?: PaymentMethodStatus | null;
  accountDirection?: AccountDirection | null;
  expiresAt?: string | null;
  createdAt?: string | null;
  updatedAt?: string | null;
};
// ...
export enum EntityClientsStatus {
  ENABLED = 'ENABLED',
  DISABLED = 'DISABLED',
  REVIEW = 'REVIEW',
  DECLINED = 'DECLINED',
}